Add ability to use event delegation
