+++ package.json
@@ -11,17 +11,17 @@
     "dev": "rollup -c scripts/rollup.dev.config.js --watch",
     "build": "npm run standalone && npm run bundle && npm run all",
     "all": "rollup -c scripts/rollup.all.config.js && rollup -c scripts/rollup.all.min.config.js",
-    "bundle": "rollup -c scripts/rollup.bundle.config.js && rollup -c scripts/rollup.bundle.min.config.js",
-    "standalone": "rollup -c scripts/rollup.standalone.config.js && rollup -c scripts/rollup.standalone.min.config.js",
+    "bundle":
+      "rollup -c scripts/rollup.bundle.config.js && rollup -c scripts/rollup.bundle.min.config.js",
+    "standalone":
+      "rollup -c scripts/rollup.standalone.config.js && rollup -c scripts/rollup.standalone.min.config.js",
     "themes": "rollup -c scripts/rollup.themes.config.js",
-    "test": "eslint 'src/**/*.js' && rollup -c scripts/rollup.tests.config.js --format es && karma start karma.conf.js",
+    "test": "eslint 'src/**/*.js'",
     "prettier": "prettier --print-width 100 --no-semi --single-quote es5 --write 'src/**/*.js'"
   },
   "author": "atomiks",
   "license": "MIT",
-  "keywords": [
-    "tooltip"
-  ],
+  "keywords": ["tooltip"],
   "devDependencies": {
     "autoprefixer": "^6.7.7",
     "babel-core": "^6.25.0",
