+++ src/js/utils/createTrigger.js
@@ -5,49 +5,45 @@ import { browser } from '../core/globals'
  * @param {String} eventType - the custom event specified in the `trigger` setting
  * @param {Element} reference
  * @param {Object} handlers - the handlers for each event
- * @param {Boolean} touchHold
+ * @param {Object} options
  * @return {Array} - array of listener objects
  */
-export default function createTrigger(eventType, reference, handlers, touchHold) {
+export default function createTrigger(eventType, reference, handlers, options) {
   const listeners = []
 
   if (eventType === 'manual') return listeners
 
-  // Show
-  reference.addEventListener(eventType, handlers.handleTrigger)
-  listeners.push({
-    event: eventType,
-    handler: handlers.handleTrigger
-  })
-
-  // Hide
-  if (eventType === 'mouseenter') {
-    if (browser.supportsTouch && touchHold) {
-      reference.addEventListener('touchstart', handlers.handleTrigger)
-      listeners.push({
-        event: 'touchstart',
-        handler: handlers.handleTrigger
-      })
-      reference.addEventListener('touchend', handlers.handleMouseleave)
-      listeners.push({
-        event: 'touchend',
-        handler: handlers.handleMouseleave
-      })
-    }
-
-    reference.addEventListener('mouseleave', handlers.handleMouseleave)
-    listeners.push({
-      event: 'mouseleave',
-      handler: handlers.handleMouseleave
-    })
+  const on = (eventType, handler) => {
+    reference.addEventListener(eventType, handler)
+    listeners.push({ event: eventType, handler })
   }
 
-  if (eventType === 'focus') {
-    reference.addEventListener('blur', handlers.handleBlur)
-    listeners.push({
-      event: 'blur',
-      handler: handlers.handleBlur
-    })
+  if (!options.target) {
+    on(eventType, handlers.handleTrigger)
+
+    if (browser.supportsTouch && options.touchHold) {
+      on('touchstart', handlers.handleTrigger)
+      on('touchend', handlers.handleMouseLeave)
+    }
+
+    if (eventType === 'mouseenter') {
+      on('mouseleave', handlers.handleMouseLeave)
+    }
+    if (eventType === 'focus') {
+      on('blur', handlers.handleBlur)
+    }
+  } else {
+    if (eventType === 'mouseenter') {
+      on('mouseover', handlers.handleDelegateShow)
+      on('mouseout', handlers.handleDelegateHide)
+    }
+    if (eventType === 'focus') {
+      on('focusin', handlers.handleDelegateShow)
+      on('focusout', handlers.handleDelegateHide)
+    }
+    if (eventType === 'click') {
+      on('click', handlers.handleDelegateShow)
+    }
   }
 
   return listeners
