+++ src/js/utils/toArray.js
@@ -1,0 +1,8 @@
+/**
+ * Ponyfill for Array.from
+ * @param {*} value
+ * @return {Array}
+ */
+export default function toArray(value) {
+  return [].slice.call(value)
+}
